requirements.txt
faiss-gpu
faiss-cpu
numpy
pandas
sentence-transformers
tqdm
transformers
FastAPI
Flask
Streamlit
openpyxl